Misplaced Delegable examples (yes/no)
Changed from 

For an example of delegable functions, consider an ID column defined as the **Number** data type in a SharePoint list. Formulas in the following example will return the results as expected. However, the first formula is non-delegable while the second is delegable.

| Formula                                           | Delegable? |
|---------------------------------------------------|------------|
| ``Filter ('SharePoint list data source', ID = 123 )`` | Yes        |
| ``Filter(`SharePoint list data source', ID ="123")``  | No         | 

to 
Changed from 

For an example of delegable functions, consider an ID column defined as the **Number** data type in a SharePoint list. Formulas in the following example will return the results as expected. However, the first formula is delegable while the second is non-delegable.

| Formula                                           | Delegable? |
|---------------------------------------------------|------------|
| ``Filter ('SharePoint list data source', ID = 123 )`` | Yes        |
| ``Filter(`SharePoint list data source', ID ="123")``  | No         |